pyproject.toml
[project]
name = "asteroid-hazard-classifier"
version = "0.1.0"
description = "Autonomous asteroid hazard classification MLOps pipeline"
authors = [
    {name = "Govinthan K S", email = "govinthanks.aiml2023@citchennai.net"}
]
license = "MIT"
requires-python = ">=3.12.1,<3.13"
dependencies = [
    "pandas>=2.2.1,<3.0.0",
    "scikit-learn>=1.4.1,<2.0.0",
    "xgboost>=2.0.3,<3.0.0",
    "fastapi>=0.110.0,<1.0.0",
    "uvicorn>=0.29.0,<1.0.0",
    "mlflow>=2.11.2,<3.0.0",
    "evidently>=0.6.0,<2.0.0",
    "great-expectations>=1.0.0,<2.0.0",
    "hydra-core>=1.3.2,<2.0.0",
    "pydantic>=2.6.4,<3.0.0",
    "gradio>=4.24.0,<5.0.0",
    "streamlit>=1.32.2,<2.0.0",
    "loguru>=0.7.2,<1.0.0",
    "requests>=2.31.0,<3.0.0",
    "python-dotenv (>=1.2.2,<2.0.0)"
]

[tool.poetry]
package-mode = false

[tool.poetry.group.dev.dependencies]
pytest = ">=8.1.1,<9.0.0"
black = ">=24.3.0,<25.0.0"
ruff = ">=0.3.4,<1.0.0"
mypy = ">=1.9.0,<2.0.0"
pre-commit = ">=3.6.2,<4.0.0"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
pythonpath = ["src"]